+++ .github/workflows/build.yml
@@ -24,6 +24,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install flake8 isort cpplint black pytest codespell h5py pylint
+        pip install annoy nmslib
         pip install -r requirements.txt
     - name: Lint with flake8
       run: |
